Refactor multithreaded token refresh code
My original intent was to use an interrupt-based approach
so a copy of the token could be made and used solely on one single thread
(i.e. the main thread)

However, such an approach would be impossible/infeasible
due to how the language works.
The new method now requires either the manual refreshing of the token
or a polling-based approach where the current state of the token
is grabbed from a worker thread that automatically refreshes periodically